Move remaining unit tests to ActionCreatorTest
