+++ src/test/kotlin/co/moelten/splity/injection/FakeConfigModule.kt
@@ -5,6 +5,8 @@ import co.moelten.splity.fakeConfig
 import me.tatarka.inject.annotations.Component
 
 @Component
-abstract class FakeConfigModule : ConfigModule {
-  override fun config(): Config = fakeConfig
+abstract class FakeConfigModule(
+  val config: Config = fakeConfig
+) : ConfigModule {
+  override fun config(): Config = config
 }
